move hnsw code into v2 package
