+++ src/java/org/apache/cassandra/index/sai/disk/vector/OnDiskOrdinalsMap.java
@@ -26,7 +26,7 @@ import com.google.common.base.Preconditions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import org.apache.cassandra.index.sai.disk.vector.hnsw.DiskBinarySearch;
+import org.apache.cassandra.index.sai.disk.v2.hnsw.DiskBinarySearch;
 import org.apache.cassandra.io.util.FileHandle;
 import org.apache.cassandra.io.util.RandomAccessReader;
 import com.github.jbellis.jvector.util.Bits;
