+++ test/unit/org/apache/cassandra/index/sai/disk/v2/hnsw/FullyConnectedHnswGraph.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.cassandra.index.sai.disk.vector.hnsw;
+package org.apache.cassandra.index.sai.disk.v2.hnsw;
 
 import java.io.IOException;
 import java.util.HashMap;
@@ -24,7 +24,7 @@ import java.util.Iterator;
 import java.util.Map;
 import java.util.List;
 
-import org.apache.cassandra.index.sai.disk.vector.hnsw.OnDiskHnswGraph.AbstractNodesIterator;
+import org.apache.cassandra.index.sai.disk.v2.hnsw.OnDiskHnswGraph.AbstractNodesIterator;
 import org.apache.cassandra.utils.Pair;
 
 import static org.apache.lucene.search.DocIdSetIterator.NO_MORE_DOCS;
