+++ test/unit/org/apache/cassandra/index/sai/disk/v2/hnsw/RandomlyConnectedHnswGraph.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.cassandra.index.sai.disk.vector.hnsw;
+package org.apache.cassandra.index.sai.disk.v2.hnsw;
 
 import java.io.IOException;
 import java.util.ArrayList;
